Don't require `and_then` to always heap-allocate
Even so that's probably going be the common-case, we
don't want to force users. Because, maybe, rather
than heap-allocating, they'd have a closed set of decoders
as a sum-type that they would prefer to inline instead
of box.

Therefore, this patch lets user in charge of deciding
what's the best allocation strategy for them.